src/requirements.txt
python-dotenv
pytest
requests
cryptography
rsa
boto3
mangum
fastapi
starlette
pydantic
pydantic-settings
uvicorn
gunicorn
Jinja2
PyJWT
sqlalchemy
sqlmodel
psycopg2
